actually, lemme just make it even better. crazy how just asking solves so many issues

+++ src/main/java/dev/cammiescorner/arcanus/common/components/entity/AuraFadeComponent.java
@@ -25,12 +25,12 @@ public class AuraFadeComponent implements AutoSyncedComponent, ServerTickingComp
 
 	@Override
 	public void readFromNbt(NbtCompound tag) {
-		timer = tag.getInt("AuraFadeTimer");
+		timer = tag.getShort("AuraFadeTimer");
 	}
 
 	@Override
 	public void writeToNbt(NbtCompound tag) {
-		tag.putInt("AuraFadeTimer", timer);
+		tag.putShort("AuraFadeTimer", (short) timer);
 	}
 
 	public int getTimer() {
